Reject thaat are obviously erroneous.
Requests under 25ms are more likely to be just before the load balancer booted up and was returning 503s and not the actual webpages.

Requests > 600 ms are rejected because they generally only count 1 or 2, but distort the scale of the plot making the rest of the data hard to understand.